Fix invalid value 'code' in ui-interactions.yml
When running vale on a tutorial, I got an error that "code" was an invalid key in styles/base/ui-interactions.yml. Reading through the Vale docs on the "existence" plugin, I saw that the proper syntax is to use the key "scope" with a value of "text" when you only want to check text and not code blocks. See https://vale.sh/docs/topics/styles/